sql/pure: restructure some PG source checks
